+++ README.md
@@ -1,2 +1,3 @@
 # java-filmorate
 Template repository for Filmorate project.
+![This is an image](src/main/resources/chart.png)
